[858] Remove dependencies to Spring Security from the event processors
All our representation event processor are connected to Spring Security
while we have no reason to do so. All those dependencies have now been
removed and the code has been simplified.

The SubscriptionManager is now in charge of the canBeDisposedSink of all
the representation event processor which removes both a field and a method
from all the representation event processor

Bug: https://github.com/eclipse-sirius/sirius-components/issues/858
