+++ backend/sirius-web-spring-collaborative-selection/src/main/java/org/eclipse/sirius/web/spring/collaborative/selection/SelectionEventProcessor.java
@@ -13,7 +13,6 @@
 package org.eclipse.sirius.web.spring.collaborative.selection;
 
 import java.util.Objects;
-import java.util.UUID;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.eclipse.sirius.web.core.api.IEditingContext;
@@ -35,7 +34,6 @@ import org.eclipse.sirius.web.spring.collaborative.selection.api.ISelectionEvent
 import org.eclipse.sirius.web.spring.collaborative.selection.dto.SelectionRefreshedEventPayload;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.core.context.SecurityContextHolder;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -67,8 +65,6 @@ public class SelectionEventProcessor implements ISelectionEventProcessor {
     private final ISubscriptionManager subscriptionManager;
 
     private final Many<IPayload> sink = Sinks.many().multicast().directBestEffort();
-
-    private final Many<Boolean> canBeDisposedSink = Sinks.many().unicast().onBackpressureBuffer();
 
     private final AtomicReference<Selection> currentSelection = new AtomicReference<>();
 
@@ -151,31 +147,8 @@ public class SelectionEventProcessor implements ISelectionEventProcessor {
         return Flux.merge(
             refreshEventFlux,
             this.subscriptionManager.getFlux(input)
-        )
-        .doOnSubscribe(subscription -> {
-            String username = SecurityContextHolder.getContext().getAuthentication().getName();
-            this.subscriptionManager.add(input, username);
-            this.logger.trace("{} has subscribed to the selection {} {}", username, this.id, this.subscriptionManager); //$NON-NLS-1$
-        })
-        .doOnCancel(() -> {
-            String username = SecurityContextHolder.getContext().getAuthentication().getName();
-            this.subscriptionManager.remove(UUID.randomUUID(), username);
-            this.logger.trace("{} has unsubscribed from the selection {} {}", username, this.id, this.subscriptionManager); //$NON-NLS-1$
-
-            if (this.subscriptionManager.isEmpty()) {
-                EmitResult emitResult = this.canBeDisposedSink.tryEmitNext(Boolean.TRUE);
-                if (emitResult.isFailure()) {
-                    String pattern = "An error has occurred while emitting that the processor can be disposed: {}"; //$NON-NLS-1$
-                    this.logger.warn(pattern, emitResult);
-                }
-            }
-        });
+        );
         // @formatter:on
-    }
-
-    @Override
-    public Flux<Boolean> canBeDisposed() {
-        return this.canBeDisposedSink.asFlux();
     }
 
     @Override
