+++ backend/sirius-web-spring-collaborative/src/main/java/org/eclipse/sirius/web/spring/collaborative/api/ISubscriptionManager.java
@@ -12,12 +12,8 @@
  *******************************************************************************/
 package org.eclipse.sirius.web.spring.collaborative.api;
 
-import java.util.List;
-import java.util.UUID;
-
 import org.eclipse.sirius.web.core.api.IInput;
 import org.eclipse.sirius.web.core.api.IPayload;
-import org.eclipse.sirius.web.spring.collaborative.dto.Subscriber;
 
 import reactor.core.publisher.Flux;
 
@@ -27,15 +23,10 @@ import reactor.core.publisher.Flux;
  * @author sbegaudeau
  */
 public interface ISubscriptionManager {
-    void add(IInput input, String username);
-
-    void remove(UUID correlationId, String username);
-
-    boolean isEmpty();
-
-    List<Subscriber> getSubscribers();
 
     Flux<IPayload> getFlux(IInput input);
+
+    Flux<Boolean> canBeDisposed();
 
     void dispose();
 
@@ -47,25 +38,12 @@ public interface ISubscriptionManager {
     class NoOp implements ISubscriptionManager {
 
         @Override
-        public void add(IInput input, String username) {
-        }
-
-        @Override
-        public void remove(UUID correlationId, String username) {
-        }
-
-        @Override
-        public boolean isEmpty() {
-            return false;
-        }
-
-        @Override
-        public List<Subscriber> getSubscribers() {
-            return List.of();
-        }
-
-        @Override
         public Flux<IPayload> getFlux(IInput input) {
+            return Flux.empty();
+        }
+
+        @Override
+        public Flux<Boolean> canBeDisposed() {
             return Flux.empty();
         }
